moved deprecation_warning to own file
